+++ packages/netlify-cms-core/src/components/Collection/FilterControl.js
@@ -23,7 +23,7 @@ const FilterControl = ({ viewFilters, t, onFilterClick, filter }) => {
   return (
     <Dropdown
       renderButton={() => {
-        return <FilterButton>{t('collection.collectionTop.filter')}</FilterButton>;
+        return <FilterButton>{t('collection.collectionTop.filterBy')}</FilterButton>;
       }}
       closeOnSelection={false}
       dropdownTopOverlap="30px"
@@ -32,11 +32,11 @@ const FilterControl = ({ viewFilters, t, onFilterClick, filter }) => {
       {viewFilters.map(viewFilter => {
         return (
           <DropdownCheckedItem
-            key={viewFilter.field}
+            key={viewFilter.id}
             label={viewFilter.label}
-            id={viewFilter.field}
-            checked={filter.get(viewFilter.field, false)}
-            onClick={() => onFilterClick(viewFilter.field)}
+            id={viewFilter.id}
+            checked={filter.getIn([viewFilter.id, 'active'], false)}
+            onClick={() => onFilterClick(viewFilter)}
           />
         );
       })}